fix: rerender in maps
